[code] XCSP3: lex constraint allows constants ; element allows list of int with a condition ; CFN: dump CFN using valid names ; pytoulbar2: linear constraint allows multiple occurences of the same variable

+++ src/xmlcsp3/XCSP3Tree.h
@@ -56,7 +56,7 @@ namespace XCSP3Core {
         Node *fromStringToTree(std::string);
 
         int arity() {
-            return listOfVariables.size();
+            return static_cast<int>(listOfVariables.size());
         }
